sync

+++ scripts/compute_clustering.py
@@ -8,14 +8,21 @@ from pathlib import Path
 from src.dataset.dataset import EventDataset
 import numpy as np
 import hdbscan
-
-filename = get_path("/work/gkrzmanc/jetclustering/results/train/Test_betaPt_BC_2025_01_03_15_07_14/eval_0.pkl", "results")
-# for rinv=0.7, see /work/gkrzmanc/jetclustering/results/train/Test_betaPt_BC_rinv07_2025_01_03_15_38_58
-
-result = CPU_Unpickler(open(filename, "rb")).load()
-dataset = EventDataset.from_directory(result["filename"], mmap=True)
-
+import argparse
 from tqdm import tqdm
+
+#filename = get_path("/work/gkrzmanc/jetclustering/results/train/Test_betaPt_BC_2025_01_03_15_07_14/eval_0.pkl", "results")
+# for rinv=0.7, see /work/gkrzmanc/jetclustering/results/train/Test_betaPt_BC_rinv07_2025_01_03_15_38_58
+# keeping the clustering script here for now, so that it's separated from the GPU-heavy tasks like inference (clustering may be changed frequently...)
+
+parser = argparse.ArgumentParser()
+parser.add_argument("--input", type=str, required=True)
+
+args = parser.parse_args()
+path = get_path(args.input, "results")
+
+#dir_results = get_path("/work/gkrzmanc/jetclustering/results/train/Test_betaPt_BC_2025_01_03_15_07_14/eval_0.pkl", "results")
+
 def get_clustering_labels(coords, batch_idx, min_cluster_size=10, min_samples=20):
     labels = []
     for i in tqdm(np.unique(batch_idx)):
@@ -26,9 +33,17 @@ def get_clustering_labels(coords, batch_idx, min_cluster_size=10, min_samples=20
         labels.append(cluster_labels)
     return np.concatenate(labels)
 
-labels = get_clustering_labels(result["pred"][:, :3], result["event_idx"])
-labels_path = os.path.join(os.path.dirname(filename), "HDBSCAN_10_20.pkl")
-
-with open(labels_path, "wb") as f:
-    pickle.dump(labels, f)
-
+for file in os.listdir(path):
+    if file.startswith("eval_") and file.endswith(".pkl"):
+        print("Computing clusters for file", file)
+        result = CPU_Unpickler(open(os.path.join(path, file), "rb")).load()
+        file_number = file.split("_")[1].split(".")[0]
+        labels_path = os.path.join(path, "clustering_{}.pkl".format(file_number))
+        if not os.path.exists(labels_path):
+            #dataset = EventDataset.from_directory(result["filename"], mmap=True)
+            labels = get_clustering_labels(result["pred"][:, :3], result["event_idx"])
+            with open(labels_path, "wb") as f:
+                pickle.dump(labels, f)
+            print("Saved labels to", labels_path)
+        else:
+            print("Labels already exist for this file")
